refactor: move getDataFromApp to a separate function
